remove spatialpandas dependency
- I included spatial pandas to provide the option to return polygons as
a spatialpandas.GeoDataFrame object that holoviz can plot directly.
This also allowed me to include a context manager to ignore shapely
deprecation warnings triggered in holoviz once here, instead of having
to create such a context manager in every piece of code that tries to
plot these geometries using holoviews.

However, requiring spatialpandas brings in as dependencies the entire
holoviz ecosystem, and this raised the time to complete conda-build from
a few minutes to about eight hours.  This tradeoff isn't worth it, so
I'm removing the dependency.