implement option to return coastlines without Chatham Islands and Kermadec islands
The Chatham Islands and the Kermadec Islands are east of 180 degress
longitude. Including these islands in a plot of New Zealand causes the
plot's horizontal axis to span -177 deg E to 177 deg E, i.e. the whole
world. An option to return the coastlines only west of 180 deg longitude
produces more readable plots for situations where the Chathams and
Kermadecs are not involved.

closes #1