test publishing to pypi from local build using twine
- just discovered that pypi package from github workflow cannot access
the data.  When installed from either the tarball or wheel from the
local build, it *can* access the data.  I'm installing to pypi
from the local build to test if the github workflow is the failure point
that prevents it accessing the data.  To do that I need to bump the version.